feat(itinerary-body): export Styled classes and nested components
re #65
