+++ packages/itinerary-body/src/defaults/index.js
@@ -1,0 +1,13 @@
+import LineColumnContent from "./line-column-content";
+import PlaceName from "./place-name";
+import RouteDescription from "./route-description";
+import TimeColumnContent from "./time-column-content";
+import TransitLegSummary from "./transit-leg-summary";
+
+export {
+  LineColumnContent,
+  PlaceName,
+  RouteDescription,
+  TimeColumnContent,
+  TransitLegSummary
+};
