+++ packages/itinerary-body/src/otp-react-redux/index.js
@@ -1,0 +1,15 @@
+import ItineraryBody from "./itinerary-body";
+import LineColumnContent from "./line-column-content";
+import PlaceName from "./place-name";
+import RouteDescription from "./route-description";
+import TransitLegSubheader from "./transit-leg-subheader";
+import ViewStopButton from "./view-stop-button";
+
+export {
+  ItineraryBody,
+  LineColumnContent,
+  PlaceName,
+  RouteDescription,
+  TransitLegSubheader,
+  ViewStopButton
+};
